fix: bounce qbittorrent when it disconnects
this has been kinda annoying me for a while now. now it should constantly check
if it is connected. when it disconnects, the pod should be bounced. this should
cause it to reconnect so i shouldnt ever be in a situation where my torrents
arent working because qbittorrent silently disconnected

Signed-off-by: Matt Provost <mattprovost6@gmail.com>